+++ include/sndfilter/wav.h
@@ -9,7 +9,7 @@
 #ifndef SNDFILTER_WAV__H
 #define SNDFILTER_WAV__H
 
-#include "snd.h"
+#include "sndfilter/snd.h"
 
 sf_snd sf_wavload(const char *file);
 bool   sf_wavsave(sf_snd snd, const char *file);
